stage 2
If a user enters a task name to the input field and presses the button with the #add-task-button id, the program should wrap the task in <li> tags and add it to the end of the <ul> tag.

The task should have the same elements as in the previous stage:

    <input> element of the checkbox type. The checkbox should be unchecked by default.
    <span> element with the task class. It contains the task name.
    <button> element with the delete-btn class. Remove the task from the list by clicking on this button.

Once the task has been added, clear the input field so that a user can add another task.

A task must have a name to be successfully added to a task list.

Also, do not remove the hardcoded tasks from the previous stage.