stage 3
A good To-Do List app can mark user's tasks as complete. Let's add this feature.

Objectives
A task should be marked as complete when a user clicks on the task checkbox. When the checkbox is checked, the <span> element with the task class should have the following properties:

text-decoration: line-through;
If the checkbox is unchecked, remove these properties from the <span> element.