Stage 4/4: Keep in memory
Objectives
We need to make our program fetch the to-do list from local storage and display the tasks with their states. Also, we need to make the program update the task list in local storage in the following cases:

A task is created.
A task is deleted.
A task is marked as complete.
We don't need the hardcoded tasks anymore, so you can delete them. When users reload the page, the program should store the tasks.